Separate Redis driver (tests are passing)
