+++ setups/saivs_client_server.json
@@ -7,6 +7,15 @@
     "sku": null,
     "mode": "client-server",
     "sai_server_ip": "172.17.0.3",
+    "driver": {
+        "type": "redis",
+        "config": {
+            "ip": "172.17.0.3",
+            "port": "6379",
+            "loglevel": "DEBUG",
+            "type": "vs"
+        }
+    },
     "port_groups": [{"1x10G": "Ethernet0", "init": "1x10G", "alias": 0},
                     {"1x10G": "Ethernet1", "init": "1x10G", "alias": 1}
                   ],
