+++ tests/ut/test_bridge_ut.py
@@ -1,5 +1,5 @@
 import pytest
-from sai import SaiObjType
+from sai_data import SaiObjType
 
 
 bport_attrs = [
